requirements.txt
keyboard==0.13.5
Pillow==9.0.1
PyAutoGUI==0.9.53
pynput==1.7.6
pytesseract==0.3.10